Fix: Fixed Font

+++ HunterPie/Update/Presentation/UpdateViewModel.cs
@@ -6,12 +6,12 @@ namespace HunterPie.Update.Presentation
     {
         private long _downloadedBytes;
         private long _totalBytes;
-        private float _downloadPercentage;
+        private double _downloadPercentage;
         private string _state;
 
         public long DownloadedBytes { get => _downloadedBytes; set { SetValue(ref _downloadedBytes, value); } }
         public long TotalBytes { get => _totalBytes; set { SetValue(ref _totalBytes, value); } }
-        public float DownloadedPercentage { get => _downloadPercentage; set { SetValue(ref _downloadPercentage, value); } }
+        public double DownloadedPercentage { get => _downloadPercentage; set { SetValue(ref _downloadPercentage, value); } }
         public string State { get => _state; set { SetValue(ref _state, value); } }
     }
 }
